\name{combine-methods}
\docType{methods}
\alias{combine-methods}
\alias{combine,AffyBatch,AffyBatch-method}
\alias{combine,AnnotatedDataFrame,AnnotatedDataFrame-method}
\alias{combine,ANY,missing-method}
\alias{combine,AssayData,AssayData-method}
\alias{combine,data.frame,data.frame-method}
\alias{combine,eSet,eSet-method}
\alias{combine,matrix,matrix-method}
\alias{combine,MIAME,MIAME-method}
\title{Methods for Function combine in Package `Biobase'}
\description{
 Methods for function \code{combine} in Package `Biobase'
}
\section{Methods}{
\describe{

\item{x = "AffyBatch", y = "AffyBatch"}{ combine AffyBatch objects of
  the same chip type}

\item{x = "AnnotatedDataFrame", y = "AnnotatedDataFrame"}{ ~~describe this method here }

\item{x = "ANY", y = "missing"}{ ~~describe this method here }

\item{x = "AssayData", y = "AssayData"}{ ~~describe this method here }

\item{x = "data.frame", y = "data.frame"}{ ~~describe this method here }

\item{x = "eSet", y = "eSet"}{ ~~describe this method here }

\item{x = "matrix", y = "matrix"}{ ~~describe this method here }

\item{x = "MIAME", y = "MIAME"}{ ~~describe this method here }
}}
\keyword{methods}
