\name{feparams-methods}
\docType{methods}
\alias{feparams-methods}
\alias{feparams,AgilentSet-method}
\title{ Get fe params for an Agilent data object}
\description{
  Get fe params for an Agilent data object}
\section{Methods}{
\describe{

\item{.Object = "AgilentSet"}{ Access the data.frame that represents the
fe params from agilent arrays--one row per sample}
}}
\keyword{methods}
