\name{festats}
\Rdversion{1.1}
\alias{festats}
\title{
Accessor for festats slot
}
\description{
Accessor for festats slot
}
\usage{
festats(.Object)
}
\arguments{
  \item{.Object}{
    An object with an festats slot.
}
}

\value{
  A data.frame with number of rows equal to the number of samples.
}
\author{
  Sean Davis <sdavis2@mail.nih.gov>
}
\seealso{
  \code{\linkS4class{AgilentSet}}
}
\examples{
}
\keyword{ methods }
