\name{Rpressa-package}
\Rdversion{1.1}
\alias{Rpressa-package}
\alias{Rpressa}
\docType{package}
\title{
  A collection of code for bioinformatics
}
\description{
  A collection of code for bioinformatics
}
\details{
\tabular{ll}{
Package: \tab Rpressa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-04-14\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}
Using the package is adhoc.
}
\author{
  Sean Davis <sdavis2@mail.nih.gov>
}
\keyword{ package }
\seealso{
}
\examples{
}
